changed hash type from intptr_t to uintptr_t; added amp_to_uint8; added amp_fatal

+++ include/amp/symbol.h
@@ -27,7 +27,7 @@
 typedef struct {
   AMP_HEAD;
   size_t size;
-  int hash;
+  uintptr_t hash;
   char *name;
 } amp_symbol_t;
